requirements.txt
aniso8601==4.1.0
asn1crypto==0.24.0
attr==0.3.1
attrs==19.1.0
Automat==0.7.0
certifi==2018.11.29
cffi==1.12.3
chardet==3.0.4
Click==7.0
constantly==15.1.0
cryptography==2.6.1
cssselect==1.0.3
docopt==0.6.2
get==2018.11.19
gevent==1.4.0
greenlet==0.4.15
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.10
locust==0.0
locustio==0.11.0
lxml==4.3.0
MarkupSafe==1.1.0
msgpack==0.6.1
Naked==0.1.31
nose==1.3.7
nose-htmloutput==0.6.0
nose-testconfig==0.10
parameterized==0.6.1
parsel==1.5.1
pipreqs==0.4.9
post==2018.11.20
public==2018.11.20
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
pycryptodome==3.7.3
pyDes==2.0.1
PyDispatcher==2.0.5
Pygments==2.3.1
PyHamcrest==1.9.0
pymongo==3.8.0
pymssql==2.1.4
PyMySQL==0.9.3
pyodbc==4.0.25
pyOpenSSL==19.0.0
pypiwin32==223
pytz==2018.9
pywin32==224
PyYAML==3.13
pyzmq==18.1.0
query-string==2018.11.20
queuelib==1.5.0
request==2018.11.20
requests==2.21.0
requests-toolbelt==0.9.1
rsa==4.0
Scrapy==1.6.0
selenium==3.141.0
service-identity==18.1.0
shellescape==3.4.1
six==1.12.0
sqlparse==0.3.0
Twisted==19.2.0
urllib3==1.24.1
w3lib==1.20.0
Werkzeug==0.14.1
yarg==0.1.9
zope.interface==4.6.0